extract to interface
