+++ beacon/sync/src/test/java/tech/pegasys/teku/beacon/sync/gossip/blocks/RecentBlocksFetchServiceTest.java
@@ -44,7 +44,7 @@ import tech.pegasys.teku.spec.datastructures.attestation.ValidatableAttestation;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 import tech.pegasys.teku.statetransition.blobs.BlockBlobSidecarsTrackersPool;
-import tech.pegasys.teku.statetransition.datacolumns.DasSamplerBasic;
+import tech.pegasys.teku.statetransition.datacolumns.DasSamplerBasicImpl;
 import tech.pegasys.teku.statetransition.util.PendingPool;
 
 public class RecentBlocksFetchServiceTest {
@@ -61,7 +61,7 @@ public class RecentBlocksFetchServiceTest {
 
   private final BlockBlobSidecarsTrackersPool blockBlobSidecarsTrackersPool =
       mock(BlockBlobSidecarsTrackersPool.class);
-  private final DasSamplerBasic dasSamplerBasic = mock(DasSamplerBasic.class);
+  private final DasSamplerBasicImpl dasSamplerBasic = mock(DasSamplerBasicImpl.class);
 
   private final FetchTaskFactory fetchTaskFactory = mock(FetchTaskFactory.class);
 
@@ -84,7 +84,7 @@ public class RecentBlocksFetchServiceTest {
             pendingBlockPool,
             pendingAttestationsPool,
             blockBlobSidecarsTrackersPool,
-            Optional.of(dasSamplerBasic),
+            dasSamplerBasic,
             forwardSync,
             fetchTaskFactory,
             maxConcurrentRequests);
